fix: removed service and tread document-service as external service
