+++ libs/clients/document-provider/README.md
@@ -4,6 +4,9 @@
 
 This library implements a client to use DocumentProvider APIs
 
+> NOTE: Since there is a race condition while creating schemas in the build process, the documents-service has to be treated as external service.
+> Therefore, the yaml has to be copied from `apps/services/documents/src/openapi.yaml` and pasted into the `clientConfig.yaml`.
+
 ### Import into other NESTJS modules
 
 Add the service to your module imports:
