+++ libs/clients/document-provider/src/index.ts
@@ -1,3 +1,2 @@
 export { ClientsDocumentProviderModule } from './lib/clients-document-provider.module'
-export { ClientsDocumentProviderService } from './lib/clients-document-provider.service'
-export * from '../src/lib/types'
+export * from '../gen/fetch'
