+++ workspace.json
@@ -2318,11 +2318,10 @@
             "passWithNoTests": true
           }
         },
-        "schemas/openapi-generator": {
+        "schemas/external-openapi-generator": {
           "builder": "@nrwl/workspace:run-commands",
           "options": {
-            "outputPath": "libs/clients/document-provider/gen/fetch",
-            "command": "yarn openapi-generator -o libs/clients/document-provider/gen/fetch -i apps/services/documents/src/openapi.yaml"
+            "command": "yarn openapi-generator -o libs/clients/document-provider/gen/fetch -i libs/clients/document-provider/src/clientConfig.yaml"
           }
         }
       }
